reproducible Debian: dont (re-)install pbuilder if it's on hold

+++ update_jdn.sh
@@ -486,13 +486,16 @@ if [ -f /etc/debian_version ] ; then
 		fi
 		$UP2DATE || ( sudo cp --preserve=mode,timestamps -r hosts/$HOSTNAME/etc/apt/sources.list /etc/apt ; sudo apt-get update )
 		$UP2DATE || sudo apt-get install $DEBS $MASTERDEBS
-		case $HOSTNAME in
-			codethink*) 	$UP2DATE || sudo apt-get install -t jessie-backports pbuilder
-					;;
-			*)		$UP2DATE || sudo apt-get install -t jessie-backports \
-						pbuilder lintian || echo "this should only fail on the first install"
-					;;
-		esac
+		# dont (re-)install pbuilder if it's on hold
+		if [ ! -n "$(dpkg -l pbuilder | egrep '^hi.*pbuilder')" ] ; then
+			case $HOSTNAME in
+				codethink*) 	$UP2DATE || sudo apt-get install -t jessie-backports pbuilder
+						;;
+				*)		$UP2DATE || sudo apt-get install -t jessie-backports \
+							pbuilder lintian || echo "this should only fail on the first install"
+				;;
+			esac
+		fi
 		# we need mock from bpo to build current fedora
 		if [ "$HOSTNAME" = "profitbricks-build3-amd64" ] || [ "$HOSTNAME" = "profitbricks-build4-amd64" ] || [ "$HOSTNAME" = "jenkins" ] ; then
 			$UP2DATE || sudo apt-get install -t jessie-backports mock \
